Store unit prices as DOUBLE PRECISION
Unit prices could only hold two decimals, which is not enough for rates
quoted per kg, per m³ or per ml. Migration 033 moves the three unit-price
columns to DOUBLE PRECISION:

  articles.prix_achat_ref
  lignes_achat.prix_unitaire
  bpu_lignes.pu_ht

Scope stops there on purpose. Line totals, HT/TVA/TTC, caisse balances,
salaries and contract amounts stay DECIMAL — those are the figures that
get invoiced, paid and reconciled, and they have to stay exact. A unit
price feeds a total that is still rounded HALF_UP to two decimals, so the
added precision is used in the calculation without leaking sub-centime
values into the ledger.

The trade-off is deliberate and worth restating: DOUBLE is binary
floating point, so 0.1 is stored approximately and repeated sums drift.
Prices should not be compared for exact equality, and the type should not
be extended to columns that are summed into a balance. Widening the
DECIMAL scale would have avoided this; DOUBLE was chosen.

Consequences handled: multiplications convert once via
BigDecimal.valueOf() before computing an exact total, @DecimalMin is
replaced by @PositiveOrZero (Bean Validation does not support DecimalMin
on floating-point types), and StockArticleRepository.sumValeurStockHt now
returns Double since its product is a float, with the dashboard rounding
it where it becomes money.

Documents two findings while here. §8.2's purchase-order cycle is
implemented exactly as specified — strictly sequential, correct roles,
stock on LIVRE, caisse debit on PAYE, 422 on a skipped step or a short
caisse — all verified end to end; only the UI was missing. And
StockArticle.seuilAlerte is never written by anything, so enAlerte can
never be true; the formula is right but the threshold has no way in.

+++ src/main/java/com/buildflow/erp/domain/achats/service/AchatServiceImpl.java
@@ -87,7 +87,10 @@ public class AchatServiceImpl implements AchatService {
                 ligne.setBpuLigne(bpuLigne);
             }
 
-            BigDecimal ligneTotal = ligneReq.quantite().multiply(ligneReq.prixUnitaire()).setScale(2, RoundingMode.HALF_UP);
+            // Price is a double now; convert once for the exact total.
+            BigDecimal ligneTotal = ligneReq.quantite()
+                    .multiply(BigDecimal.valueOf(ligneReq.prixUnitaire()))
+                    .setScale(2, RoundingMode.HALF_UP);
             ligne.setTotal(ligneTotal);
 
             totalHt = totalHt.add(ligneTotal);
@@ -209,7 +212,9 @@ public class AchatServiceImpl implements AchatService {
                 .orElseThrow(() -> new ResourceNotFoundException("LigneAchat", ligneId));
 
         ligne.setPrixUnitaire(request.prixUnitaire());
-        ligne.setTotal(ligne.getQuantite().multiply(request.prixUnitaire()).setScale(2, RoundingMode.HALF_UP));
+        ligne.setTotal(ligne.getQuantite()
+                .multiply(BigDecimal.valueOf(request.prixUnitaire()))
+                .setScale(2, RoundingMode.HALF_UP));
 
         recomputeTotals(achat);
