+++ src/main/java/com/buildflow/erp/domain/stock/repository/StockArticleRepository.java
@@ -16,6 +16,10 @@ public interface StockArticleRepository extends JpaRepository<StockArticle, UUID
 
     // No Dépôts/En Travaux split yet — StockArticle isn't scoped beyond a
     // single chantier quantity (see doc gap 2.7), so this is one global total.
+    //
+    // Returns Double, not BigDecimal: prixAchatRef is DOUBLE PRECISION, so the
+    // product and its SUM come back as a float from the database. Callers round
+    // it to two decimals before presenting it as money.
     @Query("SELECT COALESCE(SUM(s.quantiteTheorique * s.article.prixAchatRef), 0) FROM StockArticle s")
-    BigDecimal sumValeurStockHt();
+    Double sumValeurStockHt();
 }
